Scenes, Official Drag/Drop, Little Draw
Alright quite a lot in this branch. I have now officially added the
Drag and Drop function to the scene (It is a new scene). I highly
encourage everyone to add their scripts to the new scene as the new
scene has all of the maps and has my draw and drag and drops inside it.
I created a draw function (No actual animation yet) But it is working,
with a different type of card drawer every time.